Added minimatch support for file.
File can now be a minimatch expression. This allows minifying directories and updating other source files with their updated paths. This is useful for angular templateUrls, it allows one to cache bust all html files and update the js files which declare them.

I also added an option matchHashed. If match hashed is undefined or true we try to replace file-HASH.ext. Otherwise we only try to replace file.ext.